+++ Changelog.md
@@ -19,6 +19,7 @@
 - Switch to full host filesystem access through flatpak. Needed because of an unfortunate change to xdg-desktop-portal.
 - Update flatpak description
 - Ask if you want to save your project when overwriting it (creating new project, loading a new project)
+- Only create mesh acceleration structures once per mesh, not per instance
 
 ## v0.9.1 &mdash; August 28th, 2026